+++ app/components/Calls.js
@@ -1,0 +1,9 @@
+import React from "react";
+import { StyleSheet, Text, View } from "react-native";
+
+export default class Calls extends React.Component {
+  render() {
+    return <View />;
+  }
+}
+module.exports = Calls;
