--- app/components/Home.js
@@ -1,21 +1,0 @@
-import React from "react";
-import { StyleSheet, Text, View } from "react-native";
-
-export default class Home extends React.Component {
-  render() {
-    return (
-      <View style={styles.container}>
-        <Text>home</Text>
-      </View>
-    );
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center"
-  }
-});
