+++ app/package.json
@@ -10,6 +10,7 @@
     "react": "16.0.0-alpha.12",
     "react-native": "https://github.com/expo/react-native/archive/sdk-20.0.0.tar.gz",
     "react-native-scrollable-tab-view": "^0.6.7",
-    "react-native-vector-icons": "^4.4.0"
+    "react-native-vector-icons": "^4.4.0",
+    "react-navigation": "^1.0.0-beta.11"
   }
 }
